requirements.txt
Flask==3.0.3
minio==7.2.7
pillow==10.3.0
transformers==4.41.2
